pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi_rag"
version = "0.1.0"
description = "FastAPI RAG chat application"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "jinja2==3.1.2",
    "python-multipart==0.0.6",
    "langchain",
    "langchain-google-genai==0.0.7",
    "langchain-chroma==0.1.0",
    "langchain-community",
    "langgraph==0.0.26",
    "pypdf==3.17.0",
    "chromadb==0.4.18",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "bandit",
    "safety"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
